Moved types to common
