+++ package.json
@@ -4,12 +4,17 @@
   "description": "Main package file for the TCF Project",
   "author": "See repository readme",
   "scripts": {
-    "setup": "npm run setup:frontend && npm run setup:backend",
+    "setup": "npm ci && npm run setup:frontend && npm run setup:backend",
     "setup:frontend": "cd frontend/ && npm ci",
     "setup:backend": "cd backend/ && npm ci && npm run build",
     "deploy:backend": "docker-compose up --build -d mongodb && cd backend/ && npm run start",
     "deploy:frontend": "cd frontend/ && npm start",
     "deploy:dev:backend": "docker-compose up --build -d mongodb && cd backend/ && npm run dev",
     "deploy:dev:frontend": "cd frontend/ && npm start"
+  },
+  "dependencies": {
+    "@types/swagger-ui-express": "^4.1.3",
+    "mongoose": "^7.2.0",
+    "swagger-ui-express": "^4.6.3"
   }
 }
